Rename function, type-checking, style conformance
